+++ CHANGELOG.md
@@ -13,7 +13,7 @@
   - `--env` option has a default value: `base`,
   - `--datahub` is changed to `--datahub-gms-uri`, `--repository` is changed to `--docker-repository-uri`.
 - `dp deploy`'s `--docker-push` is not a flag anymore and requires a Docker repository URI parameter; `--repository` got removed then.
-- `dp run` and `dp test` run `dbt deps` before actual **dbt** command.
+- `dp run` and `dp test` run `dp compile` before actual **dbt** command.
 - Functions raise exceptions instead of exiting using `sys.exit(1)`; `cli.cli()` entrypoint is expecting exception and exits only there.
 - `dp deploy` raises an exception if there is no Docker image to push or `build/config/dag` directory does not exist.
 - Rename `gcp` to `gcs` in requirements (now one should run `pip install data-pipelines-cli[gcs]`).